+++ lib/template/renderers/inner-env/u.d.ts
@@ -2,8 +2,9 @@ import { ASTNode } from '../../../core/ast/ast-node';
 import { Renderer } from '../../../core/rendering/renderer';
 import { RenderingManager } from '../../../core/rendering/rendering-manager';
 import { WooElementKind } from '../../../util/types/woo';
-export declare class OuterEnvSolutionRenderer implements Renderer {
+/** Renderer of the u (unarticulated, underlined) inner environment */
+export declare class InnerEnvURenderer implements Renderer {
     readonly kind: WooElementKind;
-    readonly abstractVariant = "solution";
+    readonly abstractVariant = "u";
     render(renderingManager: RenderingManager, astNode: ASTNode): Node;
 }