+++ lib/util/html-components/info-block.d.ts
@@ -1,4 +1,5 @@
-export declare function infoBlockComponent({ title, children, }: {
+export declare function infoBlockComponent({ title, children, titleBackgroundColor, }: {
     title: string;
     children: Node[];
+    titleBackgroundColor?: string;
 }): Node;