+++ lib/util/lists/styles.d.ts
@@ -1,0 +1,6 @@
+export declare enum InfoBlockColors {
+    Green = "var(--color-green)",
+    Blue = "var(--color-blue)",
+    Orange = "var(--color-orange)",
+    Red = "var(--color-red)"
+}